backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
nats-py==2.9.0
kubernetes==31.0.0
asyncpg==0.30.0
pydantic==2.10.4
jinja2==3.1.4
pyyaml==6.0.2
httpx==0.28.1
python-dotenv==1.0.1
boto3==1.35.99
